2015.09.17

+++ src/actoins/contact/ActionContactImport.java
@@ -12,6 +12,7 @@ import constants.Command;
 import constants.Messages;
 import constants.Parameter;
 import constants.Uri;
+import csv.CSVContactValidator;
 import csv.ContactCSVBuilder;
 import dao.DaoException;
 import domain.Contact;
@@ -31,26 +32,39 @@ public class ActionContactImport extends AbstractActoin {
     @Override
     public String exec(HttpServletRequest request, HttpServletResponse response) throws ServiceException, DaoException {
         File file = (File) request.getAttribute(Parameter.FILE_KEY);
+
+        if (request.getAttribute(Attribute.ERROR_KEY) != null) {
+            return Uri.JSP_PREFIX + Uri.IMPORT_URI + Uri.JSP_SUFFIX;
+        }
+
+        if (!CSVContactValidator.isValid(file)) {
+            request.setAttribute(Attribute.ERROR_KEY, Messages.DATA_INCORRECT_ERROR);
+            return Uri.JSP_PREFIX + Uri.IMPORT_URI + Uri.JSP_SUFFIX;
+        }
+
         List<Contact> contacts = null;
         ContactCSVBuilder builder = new ContactCSVBuilder();
         contacts = builder.getContacts(file);
         file.delete();
         ContactService service = null;
+
         if (contacts != null && !contacts.isEmpty()) {
             service = ServiceLocatorFactory.getServiceLocator().getService(ContactService.class);
             for (Contact c : contacts) {
                 Contact findContact = service.findByLogin(c.getLogin());
+
                 if (findContact == null) {
                     service.create(c);
-                    Log.info("New contact was added to database");
+                    Log.info("New contact login=" + c.getLogin() + " was added to database");
                 } else {
                     c.setId(findContact.getId());
                     service.update(c);
+                    Log.info("Contact ID=" + c.getId() + " was updated");
                 }
             }
             return ActionManagerFactory.getActionManager().execute(Command.CONTACT_COM, request, response);
         } else {
-            request.setAttribute(Attribute.ERROR_KEY, Messages.DATA_INCORRECT_ERROR);
+            request.setAttribute(Attribute.ERROR_KEY, Messages.NO_ONE_RECORD_ERROR);
             return Uri.JSP_PREFIX + Uri.IMPORT_URI + Uri.JSP_SUFFIX;
         }
     }
